Add ui and vm packages and moved activities and viewModels to them.
