+++ app/src/main/java/com/lambdaschool/congressdata/OverviewListAdapter.kt
@@ -7,6 +7,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import com.lambdaschool.congressdata.ui.DetailsActivity
 
 import java.util.ArrayList
