+++ app/src/main/java/com/lambdaschool/congressdata/vm/CongresspersonProfileViewModel.kt
@@ -1,7 +1,9 @@
-package com.lambdaschool.congressdata
+package com.lambdaschool.congressdata.vm
 
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.ViewModel
+import com.lambdaschool.congressdata.CongresspersonProfile
+import com.lambdaschool.congressdata.ProfileRepository
 
 class CongresspersonProfileViewModel(var id: String?) : ViewModel() {
     var profile: LiveData<CongresspersonProfile>? = null
